fix: scaling bullet prefab based on object scale.
Previously the bullet scaled based on the turret object scale instead of the bullet scale.

Another issue to fix: It's currently changing the bullet prefab scale, meaning changing it in one turret will change it in all. It should be changed so that it scales the instance, instead of the prefab.